+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-08-27]
+## [1.1.0 - 2026-08-28]
 ### Added
 - Dockerfile
 - Single Spa
@@ -140,6 +140,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Override fixed form fields
 - Missing fields representation for summary in view mode
 - Calendar picker for form field of type date or timestamp
+- Filter applicants and bridgeheads
 
 ### Changed
 - Display feasibility results based only on non-empty backend results